chore: move functions to not include the "browser" module in the injected script
