+++ src/api/modules/sign/animation.ts
@@ -1,4 +1,3 @@
-import { getStoreData } from "../../../utils/background";
 import arweaveLogo from "../../../assets/arweave.png";
 
 /**
@@ -37,17 +36,4 @@ export function createCoinWithAnimation() {
     arCoin.style.transform = `translate(-${pos.x}px, -${pos.y}px)`;
     arCoin.style.opacity = `${visibility / 100}`;
   }, 100);
-}
-
-/**
- * Check if the arconfetti animation is enabled
- */
-export async function arConfettiEnabled() {
-  try {
-    const storeData = await getStoreData();
-
-    return !!storeData.settings?.arConfetti;
-  } catch {
-    return true;
-  }
 }
